Upgr groovy, change class scope, make GenericMethodTest work
Sorry guys, this update is big. First of all, I moved several tool classes
from inside ToolEntry to their own files. This will make access to these
classes shorter in general but for now it has the bad effect to cause a
lot of code changes.

There is a bug fix for the DerivedFromFilenamePattern. Formerly, the
derived from class was used as the file class itself. This lead to the
fact, that the file for which the method is called, is completely
ignored. This worked because there was no access to variables of the
file itself. However, with this commit, we introduce access to the file
group index variable (which is new and explained in the next section).
The correction should only affect the DerivedFromFilenamePattern. A
testrun of QCPipeline seemed to work.

Finally, the filegroup output object should now work. At least it does
for the tests. A file in such a group is identified by either a string
index or a numeric index (which is internally treated as a string). The
tests for the configuration factory seem to work and the tests for
GenericMethod as well. Note, that the GenericMethod call for filegroup
indices is a bit different.

Hope I did not forget anything.